\name{bcfbma-package}
\alias{bcfbma-package}
\alias{bcfbma}
\docType{package}
\title{
\packageTitle{bcfbma}
}
\description{
\packageDescription{bcfbma}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bcfbma}
\packageIndices{bcfbma}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bcfbma}

Maintainer: \packageMaintainer{bcfbma}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
