% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_ints_exact_bcf_TE.R
\name{pred_ints_exact_bcf_TE}
\alias{pred_ints_exact_bcf_TE}
\title{Prediction intervals for BCF-BMA Treatment Effect estimates}
\usage{
pred_ints_exact_bcf_TE(
  object,
  l_quant,
  u_quant,
  newdata = NULL,
  num_cores = 1,
  root_alg_precision = 1e-05
)
}
\arguments{
\item{object}{bartBMA object obtained from function bartBMA}

\item{l_quant}{Lower quartile of the prediction interval.}

\item{u_quant}{Upper quartile of the prediction interval.}

\item{newdata}{Test data for which predictions are to be produced. Default = NULL. If NULL, then produces prediction intervals for training data if no test data was used in producing the bartBMA object, or produces prediction intervals for the original test data if test data was used in producing the bartBMA object.}
}
\value{
The output is a list of length one. The one element in this list is a vector of prediction intervals???
}
\description{
This function produces prediction intervals for BCF-BMA treatment effects estimates.
}
