% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preds_bcfbma_lin_alg.R
\name{preds_bcfbma_lin_alg}
\alias{preds_bcfbma_lin_alg}
\title{Treatment Effect estimates for BCF-BMA output obtained from the posterior probability weighted averaged of the posterior means for each model}
\usage{
preds_bcfbma_lin_alg(object, num_iter, newdata = NULL, trainingdata)
}
\arguments{
\item{object}{bcfBMA object obtained from function bcfBMA}

\item{num_iter}{Total number of iterations of the Gibbs sampler (including burn-in).}

\item{newdata}{Test data for which predictions are to be produced. Default = NULL. If NULL, then produces prediction intervals for training data if no test data was used in producing the bartBMA object, or produces prediction intervals for the original test data if test data was used in producing the bartBMA object.}

\item{burnin}{Number of burn-on iterations of the Gibbs sampler.}

\item{update_resids}{Option for whether to update the partial residuals in the gibbs sampler. If equal to 1, updates partial residuals, if equal to zero, does not update partial residuals. The defaullt setting is to update the partial residua;s.}
}
\value{
The output is a list of length one. The one element in this list is a vector of prediction intervals???
}
\description{
This function produces treatment effect predictions from BCF-BMA by obtaining the posterior probability weighted average of the posterior means for each model.
}
