% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcfBMA.R
\name{bcfBMA}
\alias{bcfBMA}
\title{Bayesian Causal Forest Using Bayesian Model Averaging}
\usage{
bcfBMA(x, ...)
}
\arguments{
\item{x.train}{Training data covariate matrix excluding the treatment and propensity scores.}

\item{y.train}{Training data outcome vector.}

\item{z}{Training data treatment vector. This should be a binary vector, equal to one for treated individuals.}

\item{pihat}{Training data propensity score estimates. Each column should be a vector of propensity score estimates.}

\item{a_mu}{This is a parameter that influences the variance of the terminal node values for trees that are not interacted with the treatment.}

\item{a_tau}{This is a parameter that influences the variance of the terminal node values for trees that are interacted with the treatment.}

\item{nu}{??}

\item{sigquant}{??}

\item{c}{? This determines the size of Occam's Window}

\item{pen_mu}{??}

\item{pen_tau}{??}

\item{num_cp_mu}{??}

\item{num_cp_tau}{??}

\item{x.test}{Test data covariate matrix excluding the treatment and propensity scores.}

\item{test_z}{Test data treatment vector. This should be a binary vector, equal to one for treated individuals.}

\item{test_pihat}{Test data propensity score estimates. Each column should be a vector of propensity score estimates.}

\item{ntree_control}{Number of trees that are not interacteds with the treatment.}

\item{ntree_moderate}{Number of trees that are interacted with the treatment.}

\item{alpha_mu}{Parameter in prior probability of tree node splitting.}

\item{alpha_tau}{Parameter in prior probability of tree node splitting.}

\item{beta_mu}{Parameter in prior probability of tree node splitting.}

\item{beta_tau}{Parameter in prior probability of tree node splitting.}

\item{split_rule_node}{??}

\item{gridpoint}{??}

\item{maxOWsize}{Maximum number of models to keep in Occam's window}

\item{num_splits_mu}{Maximum number of splits in a mu tree}

\item{num_splits_tau}{Maximum number of splits in a tau tree}
}
\value{
Include lots of details here.
}
\description{
Insert description
}
